chore: dont use jest, configurable bench from json inputs
