+++ yarn-project/p2p/testbench/configurations/degree-1-strict.json
@@ -1,0 +1,15 @@
+{
+  "numberOfClients": 5,
+  "maxPeerCount": 25,
+  "gossipsubFloodPublish": false,
+  "debugDisableColocationPenalty": true,
+  "debugDisableMessageValidation": true,
+  "gossipsubInterval": 700,
+  "gossipsubD": 1,
+  "gossipsubDlo": 1,
+  "gossipsubDhi": 1,
+  "gossipsubDLazy": 1,
+  "peerCheckIntervalMS": 2500,
+  "gossipsubMcacheGossip": 12,
+  "gossipsubMcacheLength": 12
+}
