+++ yarn-project/p2p/testbench/configurations/normal-degree-100-nodes.json
@@ -1,0 +1,15 @@
+{
+  "numberOfClients": 100,
+  "maxPeerCount": 110,
+  "gossipsubFloodPublish": false,
+  "debugDisableColocationPenalty": true,
+  "debugDisableMessageValidation": true,
+  "gossipsubInterval": 700,
+  "gossipsubD": 6,
+  "gossipsubDlo": 4,
+  "gossipsubDhi": 12,
+  "gossipsubDLazy": 6,
+  "peerCheckIntervalMS": 5000,
+  "gossipsubMcacheGossip": 12,
+  "gossipsubMcacheLength": 12
+}
